Separate base agent classes
